% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_file_to_snp_sites_file.R
\name{alignment_file_to_snp_sites_file}
\alias{alignment_file_to_snp_sites_file}
\title{alignment_file_to_snp_sites_file}
\usage{
alignment_file_to_snp_sites_file(
  in_file = "./tests/data/small_alignment.aln",
  vcf = 1L,
  aln = 0L,
  phylip = 0L,
  out_file = "test"
)
}
\value{

}
\description{
alignment_file_to_snp_sites_file
}
