% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aln2vcf.R
\name{aln2vcf}
\alias{aln2vcf}
\title{Convert .fasta File to .vcf File}
\usage{
aln2vcf(fname, chromosome = 1, start_at = 0)
}
\arguments{
\item{fname}{- path to the converted file}

\item{chromosome}{- the chromosome where the fasta sequence lies}

\item{start_at}{- the first nucleotide position on the chromosome}
}
\value{
if assigned returns the data.frame of the resulting vcf file.
}
\description{
Converts a file that has multiple sequences of the same length
into a basic variant call format file using
\href{https://github.com/sanger-pathogens/snp-sites}{SNP-sites} software.
Be aware that the sequences must be of same length i.e. aligned beforehand.
}
