% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_gaps.R
\name{RemoveGaps}
\alias{RemoveGaps}
\title{Remove Gap Characters in Sequences}
\usage{
RemoveGaps(myXStringSet, removeGaps = "first", processors = 1)
}
\arguments{
\item{myXStringSet}{An \code{AAStringSet}, \code{DNAStringSet}, or
\code{RNAStringSet} object containing sequences.}

\item{removeGaps}{Determines how gaps ("-" or "." characters) are removed
in the sequences.  This should be (an unambiguous abbreviation of) one of
\code{"none"}, \code{"all"} \code{"first"} or \code{"common"}.}

\item{processors}{The number of processors to use, or \code{NULL} to
automatically detect and use all available processors.}
}
\value{
An {XStringSet} of the same type as {myXStringSet}.
}
\description{
Removes gaps ("-" or "." characters) in a set of sequences,
 either deleting all positions where the first sequence has a gap,
 all gaps that occur in any of the sequences  or only those gaps shared by
  all sequences in the set.
 This function is adopted from \link[DECIPHER]{RemoveGaps} by Erik Wright.
 It adds the gaps by first sequences to the \link[DECIPHER]{RemoveGaps}
}
\details{
The {removeGaps} argument controls which gaps are removed in
{myXStringSet}.  Setting {removeGaps} to {"first"} will remove
all gaps in the first input sequence together with the same position in the
in every sequence.  For other usages see \link[DECIPHER]{RemoveGaps}.
}
\examples{
dna <- DNAStringSet(c("ACT-G-", "AC--G-"))
RemoveGaps(dna, "first")
}
\seealso{
\link[DECIPHER]{AlignSeqs}, \link[DECIPHER]{RemoveGaps}.
}
